chore(calendar): rename CalenderDay to CalendarDay
- Renamed CalenderDay component to CalendarDay in the file
  src/components/calender/calenderDay/CalenderDay.tsx.

chore(calendar): rename calenderConstants to calendarConstants

- Renamed calenderConstants file to calendarConstants in the
  src/constants directory.

chore(calendar): rename Calender to Calendar

- Renamed Calender component to Calendar in the
  src/components/calender/calender/Calender.tsx file.

chore(calendar): rename useCalenderStore to useCalendarStore

- Renamed useCalenderStore file to useCalendarStore in the
  src/store directory.

chore(calendar): rename CALENDER to CALENDAR

- Renamed CALENDER route to CALENDAR in the src/constants/routes.ts file.

chore(calendarUtils): rename getCalenderDates to getCalendarDates

- Renamed getCalenderDates file to getCalendarDates in
