--- src/components/calender/calenderDay/CalenderDay.styles.ts
@@ -1,14 +1,0 @@
-import styled from 'styled-components';
-
-export const DayContainer = styled.div`
-  display: flex;
-  align-items: center;
-  justify-content: center;
-
-  box-sizing: border-box;
-  width: 100%;
-  height: 40px;
-  margin: 0;
-
-  border: 1px solid #ccc;
-`;
