+++ src/constants/routes.ts
@@ -2,7 +2,7 @@ const ROUTES = {
   MAIN_PAGE: '/',
   SIGN_UP: '/signup',
   SIGN_IN: '/signin',
-  CALENDER: '/calender',
+  CALENDAR: '/calendar',
 };
 
 export default ROUTES;
