--- src/pages/calenderPage/CalenderPage.tsx
@@ -1,7 +1,0 @@
-import Calender from '@/components/calender/calender/Calender.tsx';
-
-const CalenderPage = () => {
-  return <Calender />;
-};
-
-export default CalenderPage;
